test: remove redundant criteria test and add criteria adaptation test for 'exists' operator
